Se agragaron validaciones cada vez que se reciben valores en campos, funcion de eliminar  y de agragar

+++ agregar_palabra.py
@@ -34,17 +34,19 @@ def verPalabras():
     palabras = cr.fetchall()
     for palabra in palabras:
         tabla.insert("", "end", values=palabra)
- 
-
 #-----------------------------------------------------------------------------------------------------------------
 #funcion para eliminar las palabras en base al ID, se crea una ventana emergente para mejor estetica
 def eliminarPalabras():#funcion para realizar la funcion de eliminar palabras
     def eliminar():#funcion para eliminar la palabra
-        id_palabra = IdEliminado.get() #se guarda el numero ingresado por el usuario
-        cr.execute('DELETE FROM palabras WHERE id = ?', (id_palabra,))#se elimina la palabra en base al ID
-        baseDeDatos.commit()#GUARDA LOS Cambios en la base de datos
-        messagebox.showinfo("Accion exitosa", f"La palabra con ID {id_palabra} fue eliminada con exito")#muestra un mensaje una vez se eliminó la palabra
-        ventanaEliminar.destroy()#Se destruye la ventana emergente una vez se elimina la palabra
+        #validacion de los campos
+        if IdEliminado.get() == "":
+            messagebox.showwarning("Advertencia", "Por favor, ingrese un ID.")
+        else:
+            id_palabra = IdEliminado.get() #se guarda el numero ingresado por el usuario
+            cr.execute('DELETE FROM palabras WHERE id = ?', (id_palabra,))#se elimina la palabra en base al ID
+            baseDeDatos.commit()#GUARDA LOS Cambios en la base de datos
+            messagebox.showinfo("Accion exitosa", f"La palabra con ID {id_palabra} fue eliminada con exito")#muestra un mensaje una vez se eliminó la palabra
+            ventanaEliminar.destroy()#Se destruye la ventana emergente una vez se elimina la palabra
 
     # Crear ventana emergente
     ventanaEliminar = Toplevel(app)
@@ -64,17 +66,22 @@ def eliminarPalabras():#funcion para realizar la funcion de eliminar palabras
 #---------------------------------------------
 # funcion para agregar la palabra
 def agregarPalabra():
-    #se almacena el valor de las varibles en las cajas
-    nuevaPalabra = str(nombrePalabra.get())
-    nuevaDescripcion = descripcion.get("1.0", "end")
-    #se reconfigura la etiqueta de cada caja
-    resultad2.configure(text=f"La palabra es {nuevaPalabra}")
-    resultado.configure(text=f"Descripcion: {nuevaDescripcion }")
-    cr.execute('''
-        INSERT INTO palabras (palabra, descripcion)
-        VALUES (?,?)''', (nuevaPalabra, nuevaDescripcion))
-    baseDeDatos.commit()
-    messagebox.showinfo("Accion exitosa", "La palabra fue agregada con exito")
+    #validacion de los campos
+    if nombrePalabra.get() == "" or descripcion.get("1.0", "end") == "\n":
+        messagebox.showwarning("Advertencia", "Por favor, complete todos los campos.")
+        return
+    else:
+        #se almacena el valor de las varibles en las cajas
+        nuevaPalabra = str(nombrePalabra.get())
+        nuevaDescripcion = descripcion.get("1.0", "end")
+        #se reconfigura la etiqueta de cada caja
+        resultad2.configure(text=f"La palabra es {nuevaPalabra}")
+        resultado.configure(text=f"Descripcion: {nuevaDescripcion }")
+        cr.execute('''
+            INSERT INTO palabras (palabra, descripcion)
+            VALUES (?,?)''', (nuevaPalabra, nuevaDescripcion))
+        baseDeDatos.commit()
+        messagebox.showinfo("Accion exitosa", "La palabra fue agregada con exito")
 #---------------------------------
 baseDeDatos=connect("palabras.db")
 #se crea el cursor
